requirements.txt
streamlit>=1.31.0
requests>=2.31.0
google-cloud-bigquery>=3.14.0
pandas>=2.1.0
python-dotenv>=1.0.0
db-dtypes>=1.2.0
pyarrow>=14.0.0
